feat: add download_filter.py

+++ download_filter.py
@@ -2,58 +2,75 @@ import os
 import pandas as pd
 import glob
 import argparse
+from huggingface_hub import snapshot_download
+
+download_hub = "HuggingFaceFW/fineweb"
+upload_hub = "Chrisneverdie/OnlySports"
+local_download_dir = "./downloads/"
+access_token = "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"
+allow_patterns_prefix = "data/"
+
+keywords = ["football", "soccer", "basketball", "baseball", "tennis", 'athlete', 'running', 'marathon', 'copa',
+            'new', 'nike', 'adidas',
+            "cricket", "rugby", "golf", "volleyball", "sports", "sport", 'Sport', 'wrestling', 'wwe', 'hockey',
+            'volleyball', 'cycling', 'swim',
+            "athletic", "league", "team", "champion", "playoff", "olympic", 'premierleague', 'laliga', 'bundesliga',
+            'seriea', 'ligue1', 'epl',
+            'laliga', 'bundesliga', 'seriea', 'ligue1', 'racing', 'nascar', 'motogp', "cup", "worldcup", "fitness",
+            "workout", "gym", "nfl",
+            "nba", 'NBA', 'NFL', 'MLB', 'NHL', 'FIFA', 'UEFA', 'NCAA', 'MMA', 'UFC', 'ufc', "mlb", "nhl", "fifa",
+            "uefa", "ncaa", 'boxing',
+            'espn', 'bleacherreport', 'mma', 'si.com', 'formula1', 'f1', 'nytimes/athletic', 'apnews.com', 'goal']
+
+
+def download_dataset(allow_patterns):
+    filepath = snapshot_download(
+        download_hub,
+        repo_type="dataset",
+        local_dir=local_download_dir,
+        allow_patterns=allow_patterns_prefix + allow_patterns + "/*")
+    return filepath
 
 
 def filter_sports_urls(df):
-    keywords = ["football", "soccer", "basketball", "baseball", "tennis", 'athlete', 'running', 'marathon', 'copa',
-                'new', 'nike', 'adidas',
-                "cricket", "rugby", "golf", "volleyball", "sports", "sport", 'Sport', 'wrestling', 'wwe', 'hockey',
-                'volleyball', 'cycling', 'swim',
-                "athletic", "league", "team", "champion", "playoff", "olympic", 'premierleague', 'laliga', 'bundesliga',
-                'seriea', 'ligue1', 'epl',
-                'laliga', 'bundesliga', 'seriea', 'ligue1', 'racing', 'nascar', 'motogp', "cup", "worldcup", "fitness",
-                "workout", "gym", "nfl",
-                "nba", 'NBA', 'NFL', 'MLB', 'NHL', 'FIFA', 'UEFA', 'NCAA', 'MMA', 'UFC', 'ufc', "mlb", "nhl", "fifa",
-                "uefa", "ncaa", 'boxing',
-                'espn', 'bleacherreport', 'mma', 'si.com', 'formula1', 'f1', 'nytimes/athletic', 'apnews.com', 'goal']
-
     sports_filter = [url for url in df['url'] if any(keyword in url for keyword in keywords)]
     df_filtered = df.loc[df['url'].isin(sports_filter)]
     return df_filtered
 
 
-def process_parquet_files(directory):
+def process_parquet_files(pattern, group_num):
+    directory = download_dataset(pattern)
     # 获取所有的.parquet文件
     parquet_files = glob.glob(os.path.join(directory, '*.parquet'))
 
-    # 将文件等分为两组
-    mid_point = (len(parquet_files) + 1) // 2
-    group1_files = parquet_files[:mid_point]
-    group2_files = parquet_files[mid_point:]
+    for i in range(0, len(parquet_files), group_num):
+        group_files = parquet_files[i:i + group_num]
 
-    # 处理第一组文件
-    df1 = pd.concat([pd.read_parquet(file) for file in group1_files], ignore_index=True)
-    df1_filtered = filter_sports_urls(df1)
-    df1_filtered.to_parquet(os.path.join(directory, '1.parquet'))
+        if group_files:
+            # 合并文件
+            df_list = [pd.read_parquet(file) for file in group_files]
+            df_merged = pd.concat(df_list, ignore_index=True)
 
-    # 删除第一组原文件
-    for file in group1_files:
-        os.remove(file)
+            # 过滤体育相关的URL
+            df_filtered = filter_sports_urls(df_merged)
 
-    # 处理第二组文件
-    if group2_files:  # 如果第二组不为空
-        df2 = pd.concat([pd.read_parquet(file) for file in group2_files], ignore_index=True)
-        df2_filtered = filter_sports_urls(df2)
-        df2_filtered.to_parquet(os.path.join(directory, '2.parquet'))
+            # 确定输出文件名
+            first_file_name = os.path.basename(group_files[0])
+            last_file_name = os.path.basename(group_files[-1])
+            output_file_name = os.path.join(directory, f"{first_file_name.rstrip(".parquet")}_{last_file_name.rstrip(".parquet")}")
 
-        # 删除第二组原文件
-        for file in group2_files:
-            os.remove(file)
+            # 保存合并和过滤后的文件
+            df_filtered.to_parquet(output_file_name)
+
+            # 删除源文件
+            for file in group_files:
+                os.remove(file)
 
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser(description="Process parquet files to filter sports URLs.")
-    parser.add_argument("directory_path", type=str, help="Path to the directory containing parquet files.")
+    parser.add_argument("pattern", type=str, help="Target pattern in the hub")
+    parser.add_argument("group_num", type=int, help="Number of files to process in a group")
     args = parser.parse_args()
 
-    process_parquet_files(args.directory_path)
+    process_parquet_files(args.pattern, args.group_num)
